@@ -3,41 +3,41 @@
   "version": "2.0.1",
   "description": "Play the game \"Set\"",
   "main": "lib/index.js",
+  "type": "module",
   "repository": {
     "type": "git",
-    "url": "https://github.com/hildjj/set-game.git"
+    "url": "git+https://github.com/hildjj/set-game.git"
   },
   "scripts": {
-    "test": "ava",
-    "coverage": "nyc --reporter=html npm test",
-    "lint": "standard",
-    "release": "npm version patch && git push --follow-tags && npm publish",
-    "dev": "npm run coverage && light-server -w 'lib/*.js test/*.js # npm run coverage # reload' -o coverage",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"
+    "test": "c8 ava",
+    "lint": "eslint ."
   },
   "author": {
     "name": "Joe Hildebrand",
     "email": "joe-github@cursive.net"
   },
   "license": "MIT",
-  "directories": {
-    "bin": "bin"
+  "bin": {
+    "set-game": "bin/set-game.js"
   },
   "dependencies": {
-    "chalk": "^4.0.0",
-    "js-combinatorics": "^0.5.2",
-    "knuth-shuffle": "^1.0.1",
-    "pug": "^2.0.4",
-    "yargs": "^15.3.1"
+    "chalk": "^5.4.1",
+    "js-combinatorics": "^2.1.2",
+    "knuth-shuffle": "^1.0.8",
+    "pug": "^3.0.3",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "ava": "^3.8.2",
-    "light-server": "github:hildjj/light-server#update-dependencies-2020-05-11",
-    "nyc": "^15.0.1",
-    "pug-cli": "^1.0.0-alpha6",
-    "standard": "^14.3.4"
+    "@cto.af/eslint-config": "5.1.7",
+    "ava": "^6.2.0",
+    "c8": "10.1.3",
+    "eslint": "9.19.0",
+    "eslint-plugin-jsdoc": "50.6.3",
+    "eslint-plugin-n": "17.15.1",
+    "js-xiterable": "0.2.2"
   },
+  "packageManager": "pnpm@9.15.4",
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   }
 }
