@@ -1,79 +1,79 @@
-"use strict"
+import {fileURLToPath} from 'node:url';
+import path from 'node:path';
+import pug from 'pug';
+import util from 'node:util';
 
-const path = require('path')
-const util = require('util')
-const pug = require('pug')
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 /**
  * A Formatted class uses one or more Pug templates to translate into
  * different formats.
  */
-class Formatted {
-  // cache
-  static #compiled = {}
+export class Formatted {
+  // Cache
+  static #compiled = {};
 
   /**
-   * The directory to find Pug template files in
+   * The directory to find Pug template files in.
    * @static
    */
-  static TEMPLATE_DIR = path.join(__dirname, '..', 'templates')
+  static TEMPLATE_DIR = path.join(__dirname, '..', 'templates');
 
   /**
-   * Reset the cache
+   * Reset the cache.
    * @static
    */
-  static reset () {
-    Formatted.#compiled = {}
+  static reset() {
+    Formatted.#compiled = {};
   }
 
   /**
-   * Execute a template of the give output type
+   * Execute a template of the give output type.
    *
-   * @param {string} outputType - e.g. "text" or "html"
-   * @param {object} [opts={}] - util.inspect options, or anything else you want
-   * @returns {string} The rendered template
+   * @param {string} outputType E.g. "text" or "html".
+   * @param {object} [opts={}] Options for util.inspect, or anything else.
+   * @returns {string} The rendered template.
    */
-  template (outputType, opts={}) {
-    const key = `${this.constructor.name.toLowerCase()}-${outputType}`
-    let js = Formatted.#compiled[key]
+  template(outputType, opts = {}) {
+    const key = `${this.constructor.name.toLowerCase()}-${outputType}`;
+    let js = Formatted.#compiled[key];
     if (!js) {
-      const nm = path.join(Formatted.TEMPLATE_DIR, `${key}.pug`)
-      js = pug.compileFile(nm, {pretty: true})
-      Formatted.#compiled[key] = js
+      const nm = path.join(Formatted.TEMPLATE_DIR, `${key}.pug`);
+      js = pug.compileFile(nm, {pretty: true});
+      Formatted.#compiled[key] = js;
     }
-    return js({obj: this, opts})
+    return js({obj: this, opts});
   }
 
   /**
-   * Render the HTML template for this object
+   * Render the HTML template for this object.
    *
-   * @param {object} [opts={}] - util.inspect options, or anything else you want
-   * @returns {string} The rendered template
+   * @param {object} [opts={}] Options for util.inspect, or anything else.
+   * @returns {string} The rendered template.
    */
-  html (opts) {
-    return this.template('html', opts)
+  html(opts) {
+    return this.template('html', opts);
   }
 
   /**
-   * Render the text template for this object
+   * Render the text template for this object.
    *
-   * @param {object} [opts={}] - util.inspect options, or anything else you want
-   * @returns {string} The rendered template
+   * @param {object} [opts={}] Options for util.inspect.
+   * @returns {string} The rendered template.
    */
-  text (opts) {
-    return this.template('text', opts)
+  text(opts) {
+    return this.template('text', opts);
   }
 
   /**
-   * Use the text template for visualizing the object in the node REPL
+   * Use the text template for visualizing the object in the node REPL.
    *
-   * @param {number} depth - depth to descend
-   * @param {object} opts - see util.inspect for options
-   * @returns
+   * @param {number} _depth Dpth to descend.
+   * @param {object} opts See util.inspect for options.
+   * @returns {string} Potentially-colorful text.
    */
-  [util.inspect.custom] (depth, opts) {
-    return this.text(opts)
+  [util.inspect.custom](_depth, opts) {
+    return this.text(opts);
   }
 }
-
-module.exports = Formatted
