@@ -1,23 +1,26 @@
-'use strict'
+import {Formatted} from '../lib/formatted.js';
+import {fileURLToPath} from 'node:url';
+import path from 'node:path';
+import test from 'ava';
+import util from 'node:util';
 
-const util = require('util')
-const test = require('ava')
-const Formatted = require('../lib/formatted')
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 test.before(t => {
-  t.context.TEMPLATE_DIR = Formatted.TEMPLATE_DIR
-  Formatted.TEMPLATE_DIR = __dirname
-})
+  t.context.TEMPLATE_DIR = Formatted.TEMPLATE_DIR;
+  Formatted.TEMPLATE_DIR = __dirname;
+});
 
 test.after.always(t => {
-  Formatted.TEMPLATE_DIR = t.context.TEMPLATE_DIR
-})
+  Formatted.TEMPLATE_DIR = t.context.TEMPLATE_DIR;
+});
 
 test('create', t => {
-  const f = new Formatted()
-  t.is(f.html(), '\n<div class="formatted"></div>')
-  t.is(f.text(), 'Formatted')
-  t.is(f.text(), 'Formatted') // check the cache
-  Formatted.reset()
-  t.is(util.inspect(f), 'Formatted')
-})
+  const f = new Formatted();
+  t.is(f.html(), '\n<div class="formatted"></div>');
+  t.is(f.text(), 'Formatted');
+  t.is(f.text(), 'Formatted'); // Check the cache
+  Formatted.reset();
+  t.is(util.inspect(f), 'Formatted');
+});
